services/gabi-ingest/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.7
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.4.0
httpx>=0.25.0
langchain>=0.1.0
langchain-community>=0.0.10
